Suppress warnings in meult, medev2, meuct, mecom2, medsm2 & meomp

+++ mecom2/include/menps/mecom2/coll/mpi/basic_mpi_coll.hpp
@@ -49,7 +49,7 @@ public:
         ,   num_bytes
         );
         
-        mi.broadcast({ ptr, num_bytes, root_proc, comm });
+        mi.broadcast({ ptr, static_cast<int>(num_bytes), root_proc, comm });
     }
     
     void untyped_allgather(
